style: Topic info moved next to username
